pyproject.toml
[project]
name = "backend-pulsex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.120.0",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.3",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
